fix: Removes custom BitArray
