+++ Projects/Server/Json/Converters/BitArrayEnumIndexConverter.cs
@@ -14,9 +14,9 @@
  *************************************************************************/
 
 using System;
+using System.Collections;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Server.Collections;
 
 namespace Server.Json;
