+++ Projects/Server/Serialization/IGenericWriter.cs
@@ -14,9 +14,10 @@
  *************************************************************************/
 
 using System;
+using System.Collections;
 using System.IO;
 using System.Net;
-using Server.Collections;
+using System.Runtime.CompilerServices;
 
 namespace Server;
 
@@ -162,7 +163,54 @@ public interface IGenericWriter
         Write(stack);
     }
 
-    void Write(BitArray bitArray);
+    public void Write(BitArray bitArray)
+    {
+        var array = Unsafe.As<BitArray, int[]>(ref bitArray); // This is so dangerous, :O
+        var bytesLength = (bitArray.Length - 1 + (1 << 3)) >>> 3;
+
+        WriteEncodedInt(bytesLength);
+
+        var extraBits = (uint)bitArray.Length & 7;
+        if (extraBits > 0)
+        {
+            // last byte is not aligned, we will directly copy one less byte
+            bytesLength -= 1;
+        }
+
+        int quotient = Math.DivRem(bytesLength, 4, out int remainder);
+        for (int i = 0; i < quotient; i++)
+        {
+            Write(array[i]);
+        }
+
+        switch (remainder)
+        {
+            case 3:
+                {
+                    Write((byte)array[quotient]);
+                    Write((byte)(array[quotient] >> 8));
+                    Write((byte)(array[quotient] >> 16));
+                    break;
+                }
+            case 2:
+                {
+                    Write((byte)array[quotient]);
+                    Write((byte)(array[quotient] >> 8));
+                    break;
+                }
+            case 1:
+                {
+                    Write((byte)array[quotient]);
+                    break;
+                }
+        }
+
+        if (extraBits > 0)
+        {
+            // mask the final byte
+            Write((byte)((array[quotient] >> (remainder * 8)) & ((1 << (int)extraBits) - 1)));
+        }
+    }
 
     void Write(TextDefinition def)
     {
